evals/requirements.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-dotenv>=0.5.2
python-dotenv>=1.0.0
respx>=0.21.0
datasets>=2.14.0
mypy>=1.8.0
types-requests>=2.31.0
